lオプション時の整形作業を別クラスに分割
